adjust dir constructor
