--- client.js
@@ -1,2 +1,0 @@
-var client = require('./dist/lib/client/createApp')
-module.exports = client
