+++ client/index.d.ts
@@ -1,0 +1,5 @@
+export * from '../dist/client'
+
+import createApp from '../dist/client'
+
+export default createApp
