+++ client/index.js
@@ -1,0 +1,1 @@
+module.exports = require('../dist/client')
