+++ package.json
@@ -8,15 +8,12 @@
     "test:watch": "jest --coverage --watchAll",
     "prebuild": "rimraf dist",
     "build": "tsc",
-    "postbuild": "ts-node cp.ts"
+    "postbuild": "ts-node move.ts"
   },
   "files": [
-    "dist",
-    "index.js",
-    "server.js",
-    "server.d.ts",
-    "client.js",
-    "client.d.ts"
+    "share",
+    "client",
+    "index.js"
   ],
   "repository": {
     "type": "git",
@@ -35,9 +32,10 @@
   },
   "homepage": "https://github.com/Lucifier129/create-app#readme",
   "devDependencies": {
+    "@types/fs-extra": "^8.0.0",
     "@types/jest": "^23.0.1",
     "@types/node": "^12.6.9",
-    "cp-file": "^7.0.0",
+    "fs-extra": "^8.1.0",
     "jest": "^23.0.1",
     "rimraf": "^2.6.3",
     "ts-jest": "^23.0.1",
