--- server.d.ts
@@ -1,5 +1,0 @@
-import createApp from './dist/types/server'
-
-export default createApp
-
-export * from './dist/types/server'
