--- server.js
@@ -1,2 +1,0 @@
-var server = require('./dist/lib/server/createApp')
-module.exports = server
