+++ server/index.d.ts
@@ -1,0 +1,5 @@
+export * from '../dist/server'
+
+import createApp from '../dist/server'
+
+export default createApp
