+++ src/lib/type.d.ts
