+++ src/server/createApp.ts
@@ -8,10 +8,10 @@ import {
   createMemoryHistory,
   NLWithBQ
 } from 'create-history'
-import { createMap, ReqError } from '../share/util'
+import { createMap, ReqError } from '../lib/util'
 import defaultViewEngine from './viewEngine'
-import createMatcher from '../share/createMatcher'
-import defaultAppSettings from '../share/defaultSettings'
+import createMatcher from '../lib/createMatcher'
+import defaultAppSettings from '../lib/defaultSettings'
 import createController from './createController'
 import {
   Settings,
@@ -25,7 +25,7 @@ import {
   Loader,
   Route,
   Controller
-} from '../share/type'
+} from '../lib/type'
 import {
   CreateHistoryInCA,
   CreateApp,
