+++ tsconfig.json
@@ -13,13 +13,12 @@
     "allowSyntheticDefaultImports": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
-    "declarationDir": "dist/types",
-    "outDir": "dist/lib",
+    "outDir": "dist",
     "typeRoots": [
       "node_modules/@types"
     ]
   },
   "include": [
-    "src",
+    "src"
   ]
 }
